Reworked the Delaunay triangulation and added a test. The test is not complete, but it is a start.
I also added a test directory and a test file for the Delaunay triangulation.

I will countinue to work on the testing, this was my first time seriously useing pytest. So it took a while to get used to it.

I also broke the Delaunay triangulation into smaller Classes. I think this will make it easier to test and debug.  I am also reaslizing that ironcailly the Trianguation itself, may in the end be unnecessary.

The MST has the features that I need to generate the map. I may be able to just use the MST and not the triangulation. I will have to think about this more.