Continued rework of generate_random_points.
In particular, I've realized that I was could be more specific about what I was really trying to make.  The generate_random_points module now has a class called NonOverlappingPointsGenerator.

It has a class atritube called generated_points and its all golden.
I also pushed the clean code as far as I could maybe to absurdism, but in the end I feel like the code is so much more readable and I was able to so clearly articulate what I was trying to accomplish.  In particular I really enojoied the process of extracitng out the check of a conditional into its own function with a very clear name.  I tried to make the code as readable as possible, and I hope that I succeeded.