Upgraded the Triangulation Code.
I broke out the dependcy on the specific delaunay_triangulation algorithm from the triangulator code and reworked the way that the the different modules interact. So that the triangulator is now a class that can be instantiated with a specific triangulation algorithm.

Before I continue to work on the procedural generation code I wanted to make sure that the triangulation code is in a good place.

Now that it is throughly tested and has a good interface I can start to work on the procedural generation code.